+++ FOX/recipes/psf.py
@@ -95,13 +95,13 @@ API
 .. autofunction:: extract_ligand
 
 """
-import operator
 import warnings
 from os import PathLike
 from sys import version_info
+from math import ceil, floor
 from types import MappingProxyType
 from typing import (Union, Iterable, Optional, TypeVar, Callable, Mapping, Type, Iterator,
-                    Hashable, Any, Tuple, MutableMapping, AnyStr, List)
+                    Hashable, Any, Tuple, MutableMapping, AnyStr, List, SupportsFloat)
 from itertools import chain
 from collections import abc
 
@@ -414,7 +414,9 @@ def _update_lig(ligand: Molecule, k: int, copy: bool = False) -> Molecule:
     atoms_del = ligand.atoms[k:] if k != 0 else []
     for at in atoms_del:
         ligand.delete_atom(at)
-    guess_bonds(ligand)
+
+    ligand.guess_bonds()
+    dekekulize(ligand)
     fix_bond_orders(ligand)
     return ligand
 
@@ -501,42 +503,44 @@ def _get_rddict(ligands: Iterable[Union[str, Molecule, Mol]]) -> MutableMapping[
     return rdmol_dict
 
 
-AddOrSub = Callable[[float, float], float]
-FUNC_MAP: Mapping[AddOrSub, AddOrSub] = MappingProxyType({
-    operator.__add__: operator.__sub__,
-    operator.__sub__: operator.__add__
+CeilOrFloor = Callable[[SupportsFloat], int]
+FUNC_MAP: Mapping[CeilOrFloor, CeilOrFloor] = MappingProxyType({
+    ceil: floor, floor: ceil
 })
 
 
-def guess_bonds(mol: Molecule) -> None:
-    """Modified version of :meth:`Molecule.guess_bonds`.
+def dekekulize(mol: Molecule) -> None:
+    """Convert non-integer bond orders into integers.
 
     Bond orders for "aromatic" systems are no longer set to ``1.5``,
     instead addopting the more Kekulé-esque bond orders of ``1`` and ``2``.
 
+    The implemented function is a (depth-first search based) graph-walking algorithm,
+    integerifying bond orders by alternating calls to :func:`math.ceil` and :func:`math.floor`.
+    The implication herein is that :math:`i` and :math:`i+1` are considered valid (integer) values
+    for any bond order within the :math:`[i,i+1]` interval.
+
     """
-    def dfs(atom: Atom, func: AddOrSub) -> None:
+    def dfs(atom: Atom, func: CeilOrFloor) -> None:
         """Depth-first search algorithm for fixing the fixing the bond orders."""
         for b2 in atom.bonds:
             if b2.visited:
                 continue
 
             b2.visited = True
-            b2.order = int(func(b2.order, 0.5))  # Add or substract 0.5
+            b2.order = func(b2.order)  # Add or substract
             bonds.remove(b2)
 
             atom_new = b2.atom1 if b2.atom1 is not atom else b2.atom2
             dfs(atom_new, func=FUNC_MAP[func])
 
-    mol.guess_bonds()
-
     bonds = set()
     for b in mol.bonds:
-        if hasattr(b.order, 'is_integer'):
+        if hasattr(b.order, 'is_integer'):  # This catches both float and np.float instances
             if not b.order.is_integer():
                 b.visited = False
                 bonds.add(b)
-            else:
+            else:  # A float finite with integral value
                 b.visited = True
                 b.order = int(b.order)
         else:
@@ -544,10 +548,13 @@ def guess_bonds(mol: Molecule) -> None:
 
     while bonds:
         b1 = bonds.pop()
-        b1.order = int(b1.order + 0.5)
+        delta_ceil, delta_floor = ceil(b1.order) - b1.order, floor(b1.order) - b1.order
+        func = ceil if abs(delta_ceil) < abs(delta_floor) else floor
+
+        b1.order = func(b1.order)
         b1.visited = True
-        dfs(b1.atom1, func=operator.__sub__)
-        dfs(b1.atom2, func=operator.__sub__)
+        dfs(b1.atom1, func=FUNC_MAP[func])
+        dfs(b1.atom2, func=FUNC_MAP[func])
 
     for b in mol.bonds:
-        delattr(b, 'visited')
+        del b.visited